Use fluent API to configure how dokan should behave and avoid blocking apis

+++ DokanNet/DokanOptions.cs
@@ -10,7 +10,7 @@ namespace DokanNet
     /// <seealso cref="DOKAN_OPTIONS.Options"/>
     /// \endif
     [Flags]
-    public enum DokanOptions : long
+    public enum DokanOptions : uint
     {
         /// <summary>Fixed Drive.</summary>
         FixedDrive = 0,
